pyproject.toml
[project]
name = "finrag"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "docling",
    "docling-core[chunking]",
    "fastapi>=0.115.14",
    # for BAAI/bge-m3 embedding model
    "FlagEmbedding",
    "htmlrag",
    "httpx>=0.28.1",
    "huggingface-hub",
    "judgy>=0.1.0",
    "langsmith",
    "loguru",
    "marker-pdf[full]",
    "matplotlib>=3.10.3",
    "mistralai",
    "numpy<=2.2.0",
    "nltk",
    "openai",
    "opentelemetry-api",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-exporter-otlp",
    "opentelemetry-sdk",
    "opentelemetry-semantic-conventions-ai",
    "pandas>=2.3.0",
    "peft",
    "pydantic>=2.11.7",
    "pymilvus[model]",
    "pymilvus[milvus_lite]",
    "python-dotenv>=1.1.1",
    "python-fasthtml",
    # "qdrant-client",
    "qdrant-client[fastembed-gpu]",
    "rank-bm25>=0.2.2",
    "scikit-learn>=1.7.0",
    "seaborn>=0.13.2",
    "sentence-transformers",
    "tqdm>=4.67.1",
    "uvicorn>=0.35.0",
    "WeasyPrint",
    "yfinance",
    # for some reason, including this causes uv pip install -e ".[dev]"
    # to hang. comment this out and manually install it.
    # "vllm",
    # TODO: explore phoenix for tracing
]

[project.optional-dependencies]
dev = [
    "jupyter",
    "notebook",
    "ruff>=0.11.2",               # Linting and formatting
    "pre-commit>=3.4",            # Pre-commit hooks
    "finrag[test]",
]
test = [
    "uv",
    "pytest>=7.4",       # Testing
    "pytest-cov>=4.1.0", # Test coverage
    "pytest-xdist>=3.5", # Parallel testing
    "tox>=4.11",         # (Testing with) A virtual environment (locally)
]
[tool.setuptools.packages.find]
where = ["src"]
include = ["finrag*"]
